+ Use DAFRunLoop to get the modal alert view pattern working again on iOS 7.  As we mentioned in the previous commit, __beforeCARunLoopObserver is inadvertently causing a deadlock.  This is due to the way the UIAlertView is raised.  In this case, we press a table view cell to raise the UIAlertView.  This causes a UITableViewDelegate callback to occur which notifies us of the press.  On iOS 7, this occurs in the main run loop, i.e., the run loop Apple controls at the bottom of the stack, when the __beforeCARunLoopObserver fires.  Further up the stack, we run *our* run loop in order to get a UIAlertView to behave modally.  When the user attempts to dismiss this alert view (via the “Cancel” button in our case), this event is expected to be handled on __beforeCARunLoopObserver as well.  However, __beforeCARunLoopObserver is effectively “locked up” at the bottom of the stack in a “firing” state.  Thus *our* run loop skips over it.  This behavior can be seen more explicitly in Apple’s CoreFoundation code in the __CFRunLoopDoObservers() function.
[NOTES]
This is a proof of concept which is not meant to be used in production code.  There is a reason why Apple uses a “firing” state to guard reentrancy.  That said, Apple has introduced a nasty regression that really prevents safe usage of recursive NSRunLoops, and contradicts patterns that they themselves publish in their own documentation.  From Apple’s own “Threading Programming Guide” [0]:

It is possible to run a run loop recursively. In other words, you can call CFRunLoopRun, CFRunLoopRunInMode, or any of the NSRunLoop methods for starting the run loop from within the handler routine of an input source or timer. When doing so, you can use any mode you want to run the nested run loop, including the mode in use by the outer run loop.

In fact, the example they state in listing 3-2 is the exact pattern that the modal alert view pattern uses.

[0] https://developer.apple.com/library/ios/documentation/cocoa/Conceptual/Multithreading/RunLoopManagement/RunLoopManagement.html

— daf